Adding junits for refactored UEB Listener
Changes made:
Added null check in SdncBaseModel.java
Cleaned up in SdncVFCModel.java
Added tests in SdncARModelTest.java, SdncNodeModelTest and SdncVFCModelTest
Added files SdncGroupModelTest, SdncServiceModelTest and SdncVFModelTest

Issue-ID: CCSDK-360

+++ ueb-listener/src/main/java/org/onap/ccsdk/sli/northbound/uebclient/SdncBaseModel.java
@@ -887,13 +887,15 @@ public class SdncBaseModel {
              try {
             	int rowCount = 0;
             	CachedRowSet data = jdbcDataSource.getData("SELECT * from " + tableName + " where " + keyName + " = " + keyValue + ";", null, "");
-            	while(data.next()) {
+            	if (data != null) {
+            		while(data.next()) {
      				rowCount ++;
-            	}
-            	if (rowCount != 0) {
-                    LOG.debug("cleanUpExistingToscaData from: " + tableName + " for " + keyValue);
-               		jdbcDataSource.writeData("DELETE from " + tableName + " where " + keyName + " = " + keyValue + ";", null, null);
-            	}
+            		}
+            		if (rowCount != 0) {
+                    	LOG.debug("cleanUpExistingToscaData from: " + tableName + " for " + keyValue);
+               			jdbcDataSource.writeData("DELETE from " + tableName + " where " + keyName + " = " + keyValue + ";", null, null);
+            		}
+		}
 
 			} catch (SQLException e) {
 				LOG.error("Could not clean up existing " + tableName  + " for " + keyValue, e);
@@ -907,13 +909,15 @@ public class SdncBaseModel {
             try {
            	int rowCount = 0;
            	CachedRowSet data = jdbcDataSource.getData("SELECT * from " + tableName + " where " + key1Name + " = " + key1Value + " AND " + key2Name + " = " + key2Value + ";", null, "");
-           	while(data.next()) {
+           	if (data != null) {
+           		while(data.next()) {
     				rowCount ++;
-           	}
-           	if (rowCount != 0) {
-                   LOG.debug("cleanUpExistingToscaData from : " + tableName + " for " + key1Value + " and " + key2Value);
-              		jdbcDataSource.writeData("DELETE from " + tableName + " where " + key1Name + " = " + key1Value + " AND " + key2Name + " = " + key2Value + ";", null, null);
-           	}
+           		}
+           		if (rowCount != 0) {
+                   	LOG.debug("cleanUpExistingToscaData from : " + tableName + " for " + key1Value + " and " + key2Value);
+              			jdbcDataSource.writeData("DELETE from " + tableName + " where " + key1Name + " = " + key1Value + " AND " + key2Name + " = " + key2Value + ";", null, null);
+           		}
+		}
 
 			} catch (SQLException e) {
 				LOG.error("Could not clean up existing " + tableName  + " for " + key1Value  + " and " + key2Value, e);
